[FIX] im_livechat: notify operator on leave channel for sharable support
when we are using the sharable support page (url : "/im_livechat/support/"),
There is no "visitor left" message sent to the operator when any visitor left
the channel.
Ideally, when any visitor left the channel from the website or the support page
should notify the operator that "Visitor has left the conversation".

With this commit, the operator will get notified every time when visitor left
the conversation from anywhere.

Task-2210040
